refactor: flatten tab routes
