+++ app/orders.tsx
@@ -9,14 +9,14 @@ import {
 import { useAppRouter } from '@/services';
 import { Package, ArrowLeft, ChevronLeft, ShoppingBag, Clock, Truck } from 'lucide-react-native';
 import { useAuth } from '@/features/auth/AuthContext';
-import { Order, CartItem } from '../../types';
+import { Order, CartItem } from '../types';
 import { useTheme } from '@/ui/ThemeProvider';
 import EmptyState from '@/shared/ui/EmptyState';
-import OrderTrackingModal from '../../components/OrderTrackingModal';
+import OrderTrackingModal from '../components/OrderTrackingModal';
 import { useAuthModal } from '@/features/auth/AuthModalContext';
 import commonStyles from '@/constants/styles';
 import { useLanguage } from '@/ui/ThemeProvider';
-import { useCurrency } from '../../contexts/CurrencyContext';
+import { useCurrency } from '../contexts/CurrencyContext';
 import { useOrders } from '@/services';
 import { requireEnv } from '@/services/config';
